requirements.txt
# These are the requirements from Climada 2.2.0
bottleneck>=1.3
cartopy>=0.18,<0.20.0
# cfgrib>=0.9.7
contextily>=1.0
dask[dataframe]>=2.25
# geopandas>=0.8
h5py>=2.10
haversine>=2.3
matplotlib>=3.2
netcdf4>=1.5
# numba>=0.51
numba==0.54 # Pinned because sets numpy requirement (numpy <= 1.20)
openpyxl>=3.0
pandas-datareader>=0.9
pathos>=0.2
pint>=0.15
pycountry>=20.7
pyepsg>=0.4
tables>=3.6
pyxlsb>=1.0
rasterio>=1.1,<1.2.7
requests>=2.24
# salib>=1.3.0
scikit-learn>=0.23
scipy>=1.6
statsmodels>=0.11
tabulate>=0.8
tqdm>=4.48
# xarray>=0.13
xlrd>=1.2
xlsxwriter>=1.3
xmlrunner>=1.7
# Climada pip reqs
deprecation>=2.1
overpy>=0.4
peewee>=3.14
# pybufrkit>=0.2
SALib==1.3.12

# Project specific reqs
beautifulsoup4==4.9.1
cfgrib==0.9.9.0
click==8.0.1
feedparser==6.0.8
#geopandas==0.9.0
geopandas==0.10.2
lxml==4.6.3
numpy==1.20 # Check numba req before updating
pandas==1.3.1
pybufrkit==0.2.19
Rtree==0.9.4
shapely==1.7.1
xarray==0.19.0
azure-storage-file==2.1.0
rasterstats==0.14.0
seaborn==0.11.2
azure-keyvault-secrets
azure-identity